requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18
PyPDF2==3.0.1
deep-translator==1.11.4
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
